Smart school
In oder to help schools in lebanon in many domains(Technology, Rational consumption Awareness, Avoid risks and saving money and time, An Educational Motivating Means, Solving the parking problem), using smart phone aaplication and raspberry pi, we can turn ON or OFF the genarator in school, Turn ON or OFF each room, Send Message to an electronic annoucement sign to represent the First on the school list, and Using this application allows the smart school to announce its occasions as well as its values and students achievements through a number of messages posted on this screen. Students then will be always expecting what is new in their schoo,also, we can find the empty parking from phone online.